Update for HumHub 1.14 (#2)
* Update spaceChooser.php

* Update link.php

* Update comments.php

* Update form.php

* Update showFiles.php

* Update wallCreateContentForm.php

* Update wallCreateContentForm.php

* Delete wallStreamFilterNavigation.php

It seems there were no theme specific features in this file. By removing it, the default HumHub view is taken.
The file would have to be updated with every change in HumHub.

* Update form.php

trying to keep the theme specific changes in place (not perfectly)

* Update form.php

only difference to HumHub default: maxHeight: 600px instead of 300px

* Update edit.php

* Update stream.less

* Update theme.css

* Import select2 styles

e.g. fixes the space selection field in the search filter

* Update theme.css

* Fix editing of posts

* Fix editing of comments

* Update build.less

* Update theme.css